리뷰 등록 구현 완료. 클래스 메인페이지 들어가면 각 클래스별로 평점 보여주는거 하는중

+++ src/main/java/com/knocksea/see/user/service/UserService.java
@@ -249,7 +249,7 @@ public class UserService {
                 List<Review> findByShipReviews = reviewRepository.findByProduct(product);
                 entireInfoResponseDTO.setShipReviewList(findByShipReviews);
                 for (Review findByShipReview : findByShipReviews) {
-                    Long reviewRating = findByShipReview.getReviewRating();
+                    Double reviewRating = findByShipReview.getReviewRating();
                     shipReviewTotal += reviewRating;
                 }
 
@@ -259,7 +259,7 @@ public class UserService {
                 List<Review> findBySpotReviews = reviewRepository.findByProduct(product);
                 entireInfoResponseDTO.setSpotReviewList(findBySpotReviews);
                 for (Review findBySpotReview : findBySpotReviews) {
-                    Long reviewRating = findBySpotReview.getReviewRating();
+                    Double reviewRating = findBySpotReview.getReviewRating();
                     spotReviewTotal +=reviewRating;
                 }
                 entireInfoResponseDTO.setSpotReviewAvgScore((int)spotReviewTotal/findBySpotReviews.size());
@@ -270,7 +270,7 @@ public class UserService {
             List<Review> findByEduReviews = reviewRepository.findAllByEdu(edu);
             entireInfoResponseDTO.setEduReviewList(findByEduReviews);
             for (Review findByEduReview : findByEduReviews) {
-                Long reviewRating = findByEduReview.getReviewRating();
+                Double reviewRating = findByEduReview.getReviewRating();
                 eduReviewTotal+=reviewRating;
             }
             entireInfoResponseDTO.setEduReviewAvgScore((int)eduReviewTotal/findByEduReviews.size());
